rsyslogd: Fix race between signals and main loop timeout
The main loop sleeps in a select() call for a long interval in order to
periodically run housekeeping tasks. The main loop is also responsible for
responding to flags set by signal handlers, so this sleeping should be
interrupted by signals so that it can check those flags.

However, a signal could be delivered between when the flags are checked and
when select() is called. In which case the main loop will block for the
full interval (currently 10 minutes) before handling the signal. If this
occurs, it could take up to 10 minutes for rsyslogd to terminate after a
SIGTERM or respond to SIGHUP.

Fix this by blocking signals before checking the flags and using pselect()
to unblock the signals while waiting. This is recommended by the select(2)
manual page to avoid this very issue.

Signed-off-by: John Brooks <jbrooks@ciena.com>