queue cleanup: remove no longer needed debug output
This was added for a specific debug effort and obviously forgotten
to cleanup. This issue is not included in any scheduled release. It
was added a few days ago (we did not try to hunt down the exact
commit that caused it).

Thanks to github user eshadesu for alerting us.

closes https://github.com/rsyslog/rsyslog/issues/3955