imjournal bugfix: do not wait too long on recovery try
When trying to recover journal errors, imjournal waited a hardcoded
period of 10s between tries. This was pretty long and could lead to
loss of journal data.

This commit adjust it to 100ms, which should still be fully sufficient
to prevent the journal from "hammering" the CPU.

It may be worth considering to make this setting configurable - but
let's first see if there is real demand to actually do that.

closes https://github.com/rsyslog/rsyslog/issues/3969