testbench: do not create hanging tool instances
the omhttp test server hangs indefinitely if tests requiring
it fail. This is relevant if the environment ist not destructed
after CI run (the buildbot CI system does not do this on VMs).

To guard against this, we have added a timeout of 30 minutes to
ensure the tool is terminated.